# docs/swagger.yaml
basePath: /
...
host: https://quickmailserver-production.up.railway.app
info:
  contact:
    email: karansingh122134@gmail.com
    name: Karan Singh
    url: https://github.com/karan-singh-17/Quick_Mail_server
  description: 'The Quick Mail API allows users to create groups, add recipients manually
    or from a CSV file (either from a device or a link), and send messages (plain
    text, HTML from a link, or an HTML file) to multiple recipients in a single API
    call. This API is designed to simplify mass email communications by providing
    flexible options for managing recipient lists and message content. \n ### Use
    Cases:\n1. **Group Creation and Email Campaigns**: A user can create a group and
    manually add recipients or import them from a CSV file. The user can then send
    a message to all members of the group. \n2. **Dynamic Email Campaigns**: If a
    user has a dynamic list of recipients stored online in a CSV file, they can use
    the file''s URL to add recipients to the group without manually updating the list.
    \n3. **HTML Email Campaigns**: Users can send HTML emails either by uploading
    an HTML file or by providing a link to an HTML template hosted online. \n4. **Simple
    Text Campaigns**: Users can quickly send plain text messages to a list of recipients
    without the need for HTML formatting. \n ### Security: \nThe Quick Mail API is
    secure and incorporates a custom-built two-factor authentication (2FA) system.
...
  title: Quick Mail API
...
paths:
...
  /api/group/execute-group:
    post:
      consumes:
      - application/json
      description: starts the process of sending emails to the recipients. Make sure
        you are logged in and are the owner of the group.
      parameters:
      - description: Group ID
        in: body
        name: group_id
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid Input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Group not found
          schema:
            type: string
        "500":
          description: Error in sending mails
          schema:
            type: string
      summary: execute/run the group
      tags:
      - Groups
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: This endpoint verifies user credentials by checking the email and
        password. If valid, it sends a login code to the user's email for further
        verification. The request must be a POST method with a JSON body containing
        "email" and "password". The login code is stored temporarily for verification
        purposes.
      parameters:
      - description: Email and Password
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login code sent to email
          schema:
            type: string
        "400":
          description: 'Invalid Input: Error parsing JSON body'
          schema:
            type: string
        "401":
          description: 'Unauthorized: Invalid email or password'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: Failed to send login code'
          schema:
            type: string
      summary: Authenticate user and send login code
      tags:
      - User Authentication
  /api/user/logout:
    post:
      consumes:
      - application/json
      description: This endpoint logs out the user by clearing the JWT token cookie.
        The request must be a POST method. Upon successful sign-out, the JWT token
        is removed from the cookies, and a success message is returned.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully signed out
          schema:
            type: string
        "405":
          description: 'Method Not Allowed: Invalid request method'
          schema:
            type: string
      summary: Sign out the user and clear the JWT token
      tags:
      - User Authentication
  /api/user/register:
    post:
      consumes:
      - application/json
      description: This endpoint allows a new user to register by providing an email
        and password. The email is checked for existing registration, and a verification
        token is generated and sent via email. The user must verify their email to
        complete the registration process.
      parameters:
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: User registered. Please check your email to verify your account."
            "User registration successful
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: Email is already registered
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - User
  /api/user/verify-login-code:
    post:
      consumes:
      - application/json
      description: This endpoint verifies the provided login code for the specified
        email. If the code is valid, a JWT token is generated and set as a cookie
        in the response.
      parameters:
      - description: Email and login code
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            type: string
        "400":
          description: 'Invalid Input: Error parsing request body'
          schema:
            type: string
        "401":
          description: 'Unauthorized: Invalid or expired login code'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: Failed to generate token'
          schema:
            type: string
      summary: Verify the login code and generate JWT token
      tags:
      - User Authentication
...
swagger: "2.0"